+++ CHANGELOG.md
@@ -8,6 +8,9 @@ This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachang
 ### Changed
 - Mesos check supports multiple servers.
 
+### Added
+- Basic chronos check.
+
 ## [0.0.2] - 2015-07-14
 
 ### Changed